Extract `UnixListener` class to a dedicated file
A later commit will consume the `UnixListener` class in another test file,
so move it to a shared one.
